Updated almost everything
Almost every method has been updated to integrate the new Int class, as well as both positive and negative integers now. The only method that has not is the exponent method, since in order to allow for negative exponents I will need to incorporate division, which before doing that will probably require incorporating decimals.

More to come soon for sure.